[Brendon] - GameObject model update
Removed width and height attributes. If needed, you can access the active sprite width & height attributes.

Added support for GameObject's with a static sprite, for objects that don't need an animation.

Added a new load() method and changed the old init().
Now init() is called when the object is initialized and load is called when the state in which the object is present is loaded.